tonemapping pass, code restructuring, more input controls
