+++ README.md
@@ -9,14 +9,35 @@ other helpers.
 * [x] Physically based shading model
 * [x] Point lights
 * [x] Basic `glTF` import
-* [ ] HDR rendering with a tone mapping pass
+* [x] HDR rendering with a tone mapping pass
 * [ ] Diffuse and specular environment mapping
 * [ ] Directional lights
 * [ ] Shadow mapping
 * [ ] More robust `glTF` import
+
+# Controls
+
+### Navigation
+* **Left click**: Rotate camera
+* **Middle click**: Pan camera
+* **Right click/Scroll wheel**: Dolly camera
+
+### Model controls
+* **X**: Add a row of models in the X direction
+* **Y**: Add a row of models in the X direction
+* **Z**: Add a row of models in the X direction
+
+\* *Hold shift to subtract a row*
+
+### Tonemapping/Exposure controls
+* **A**: Use ACES Tonemapping curve
+* **U**: Use Uncharted 2 Tonemapping curve
+* **C**: Display Uncharted 2 and ACES in split-screen configuration
+* **Hold CTRL + left click**: Adjust split screen split
 
 # Screenshots
 
 ![](screenshots/helmet1.png)
 ![](screenshots/helmet2.png)
 ![](screenshots/helmet3.png)
+![](screenshots/helmet4.png)
