+++ src/pass/tonemap.rs
@@ -1,0 +1,346 @@
+use rendy::{
+    command::{QueueId, RenderPassEncoder},
+    factory::{Factory},
+    graph::{render::*, NodeBuffer, NodeImage, ImageAccess},
+    hal::{pso::DescriptorPool, Device},
+    resource::{buffer::Buffer, image::{ImageView, Filter, WrapMode, ViewKind}, sampler::Sampler},
+    shader::{Shader, ShaderKind, SourceLanguage, StaticShaderInfo},
+};
+
+use gfx_hal as hal;
+
+use std::mem::size_of;
+
+use crate::Aux;
+
+lazy_static::lazy_static! {
+    static ref VERTEX: StaticShaderInfo = StaticShaderInfo::new(
+        concat!(env!("CARGO_MANIFEST_DIR"), "/assets/shaders/tonemap.vert"),
+        ShaderKind::Vertex,
+        SourceLanguage::GLSL,
+        "main",
+    );
+
+    static ref FRAGMENT: StaticShaderInfo = StaticShaderInfo::new(
+        concat!(env!("CARGO_MANIFEST_DIR"), "/assets/shaders/tonemap.frag"),
+        ShaderKind::Fragment,
+        SourceLanguage::GLSL,
+        "main",
+    );
+}
+
+#[derive(Debug, Clone, Copy)]
+#[repr(C)]
+pub struct TonemapperArgs {
+    pub exposure: f32,
+    pub curve: i32,
+    pub comparison_factor: f32,
+}
+
+impl std::fmt::Display for TonemapperArgs {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "Exposure: {}, Curve: {}", self.exposure, self.curve)
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+#[repr(C)]
+pub struct UniformArgs {
+    tonemapper: TonemapperArgs,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+struct Settings {
+    align: u64,
+}
+
+impl<B: hal::Backend> From<&Aux<B>> for Settings {
+    fn from(aux: &Aux<B>) -> Self {
+        Self::from_aux(aux)
+    }
+}
+
+impl<B: hal::Backend> From<&mut Aux<B>> for Settings {
+    fn from(aux: &mut Aux<B>) -> Self {
+        Self::from_aux(aux)
+    }
+}
+
+impl Settings {
+    const UNIFORM_SIZE: u64 = size_of::<UniformArgs>() as u64;
+
+    fn from_aux<B: hal::Backend>(aux: &Aux<B>) -> Self {
+        Settings {
+            align: aux.align,
+        }
+    }
+
+    #[inline]
+    fn buffer_frame_size(&self) -> u64 {
+        ((Self::UNIFORM_SIZE - 1) / self.align + 1) * self.align
+    }
+
+    #[inline]
+    fn uniform_offset(&self, index: u64) -> u64 {
+        self.buffer_frame_size() * index as u64
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct PipelineDesc;
+
+#[derive(Debug)]
+pub struct Pipeline<B: gfx_hal::Backend> {
+    buffer: Buffer<B>,
+    sets: Vec<B::DescriptorSet>,
+    descriptor_pool: B::DescriptorPool,
+    image_sampler: Sampler<B>,
+    image_view: ImageView<B>,
+    settings: Settings,
+}
+
+impl<B> SimpleGraphicsPipelineDesc<B, Aux<B>> for PipelineDesc
+where
+    B: gfx_hal::Backend,
+{
+    type Pipeline = Pipeline<B>;
+
+    fn images(&self) -> Vec<ImageAccess> {
+        vec![
+            ImageAccess {
+                access: rendy::resource::image::Access::SHADER_READ,
+                usage: hal::image::Usage::SAMPLED,
+                layout: hal::image::Layout::ShaderReadOnlyOptimal,
+                stages: hal::pso::PipelineStage::FRAGMENT_SHADER,
+            }
+        ]
+    }
+
+    fn depth_stencil(&self) -> Option<gfx_hal::pso::DepthStencilDesc> {
+        None
+    }
+
+    fn load_shader_set<'a>(
+        &self,
+        storage: &'a mut Vec<B::ShaderModule>,
+        factory: &mut Factory<B>,
+        _aux: &mut Aux<B>,
+    ) -> gfx_hal::pso::GraphicsShaderSet<'a, B> {
+        storage.clear();
+
+        log::trace!("Load shader module '{:#?}'", *VERTEX);
+        storage.push(VERTEX.module(factory).unwrap());
+
+        log::trace!("Load shader module '{:#?}'", *FRAGMENT);
+        storage.push(FRAGMENT.module(factory).unwrap());
+
+        gfx_hal::pso::GraphicsShaderSet {
+            vertex: gfx_hal::pso::EntryPoint {
+                entry: "main",
+                module: &storage[0],
+                specialization: gfx_hal::pso::Specialization::default(),
+            },
+            fragment: Some(gfx_hal::pso::EntryPoint {
+                entry: "main",
+                module: &storage[1],
+                specialization: gfx_hal::pso::Specialization::default(),
+            }),
+            hull: None,
+            domain: None,
+            geometry: None,
+        }
+    }
+
+    fn layout(&self) -> Layout {
+        Layout {
+            sets: vec![
+                SetLayout {
+                    bindings: vec![
+                        hal::pso::DescriptorSetLayoutBinding {
+                            binding: 0,
+                            ty: hal::pso::DescriptorType::Sampler,
+                            count: 1,
+                            stage_flags: hal::pso::ShaderStageFlags::FRAGMENT,
+                            immutable_samplers: false,
+                        },
+                        hal::pso::DescriptorSetLayoutBinding {
+                            binding: 1,
+                            ty: hal::pso::DescriptorType::SampledImage,
+                            count: 1,
+                            stage_flags: hal::pso::ShaderStageFlags::FRAGMENT,
+                            immutable_samplers: false,
+                        },
+                        hal::pso::DescriptorSetLayoutBinding {
+                            binding: 2,
+                            ty: hal::pso::DescriptorType::UniformBuffer,
+                            count: 1,
+                            stage_flags: hal::pso::ShaderStageFlags::FRAGMENT,
+                            immutable_samplers: false,
+                        }
+                    ]
+                }
+            ],
+            push_constants: Vec::new(),
+        }
+    }
+
+    fn build<'a>(
+        self,
+        factory: &mut Factory<B>,
+        _queue: QueueId,
+        aux: &mut Aux<B>,
+        buffers: Vec<NodeBuffer<'a, B>>,
+        images: Vec<NodeImage<'a, B>>,
+        set_layouts: &[B::DescriptorSetLayout],
+    ) -> Result<Pipeline<B>, failure::Error> {
+        assert!(buffers.is_empty());
+        assert!(images.len() == 1);
+        assert!(set_layouts.len() == 1);
+
+        let settings: Settings = aux.into();
+
+        let frames = aux.frames;
+
+        let mut descriptor_pool = unsafe {
+            factory.create_descriptor_pool(
+                frames,
+                vec![
+                    hal::pso::DescriptorRangeDesc {
+                        ty: hal::pso::DescriptorType::Sampler,
+                        count: frames,
+                    },
+                    hal::pso::DescriptorRangeDesc {
+                        ty: hal::pso::DescriptorType::SampledImage,
+                        count: frames,
+                    },
+                    hal::pso::DescriptorRangeDesc {
+                        ty: hal::pso::DescriptorType::UniformBuffer,
+                        count: frames,
+                    },
+                ],
+            )
+        }
+        .unwrap();
+
+        let image_sampler = factory.create_sampler(Filter::Nearest, WrapMode::Clamp).unwrap();
+
+        let image_view = factory.create_image_view(
+            images[0].image,
+            ViewKind::D2,
+            hal::format::Format::Rgba32Float,
+            hal::format::Swizzle::NO,
+            images[0].range.clone(),
+        ).expect("Could not create tonemapper input image view");
+
+        let buffer = factory
+            .create_buffer(
+                aux.align,
+                settings.buffer_frame_size() * aux.frames as u64,
+                (
+                    hal::buffer::Usage::UNIFORM,
+                    rendy::memory::MemoryUsageValue::Dynamic,
+                ),
+            )
+            .unwrap();
+
+        let mut sets = Vec::with_capacity(frames);
+        for index in 0..frames {
+            unsafe {
+                let set = descriptor_pool.allocate_set(&set_layouts[0]).unwrap();
+                factory.write_descriptor_sets(vec![
+                    hal::pso::DescriptorSetWrite {
+                        set: &set,
+                        binding: 0, 
+                        array_offset: 0,
+                        descriptors: Some(hal::pso::Descriptor::Sampler(
+                            image_sampler.raw(),
+                        )),
+                    },
+                    hal::pso::DescriptorSetWrite {
+                        set: &set,
+                        binding: 1,
+                        array_offset: 0,
+                        descriptors: Some(hal::pso::Descriptor::Image(
+                            image_view.raw(),
+                            hal::image::Layout::ShaderReadOnlyOptimal,
+                        )),
+                    },
+                    hal::pso::DescriptorSetWrite {
+                        set: &set,
+                        binding: 2,
+                        array_offset: 0,
+                        descriptors: Some(hal::pso::Descriptor::Buffer(
+                            buffer.raw(),
+                            Some(settings.uniform_offset(index as u64))..Some(settings.uniform_offset(index as u64) + Settings::UNIFORM_SIZE),
+                        )),
+                    },
+                ]);
+                sets.push(set);
+            }
+        }
+        Ok(Pipeline {
+            buffer,
+            sets,
+            image_view,
+            image_sampler,
+            descriptor_pool,
+            settings,
+        })
+    }
+}
+
+impl<B> SimpleGraphicsPipeline<B, Aux<B>> for Pipeline<B>
+where
+    B: gfx_hal::Backend,
+{
+    type Desc = PipelineDesc;
+
+    fn prepare(
+        &mut self,
+        factory: &Factory<B>,
+        _queue: QueueId,
+        _set_layouts: &[B::DescriptorSetLayout],
+        index: usize,
+        aux: &Aux<B>,
+    ) -> PrepareResult {
+        unsafe {
+            factory
+                .upload_visible_buffer(
+                    &mut self.buffer,
+                    self.settings.uniform_offset(index as u64),
+                    &[UniformArgs {
+                        tonemapper: aux.tonemapper_args,
+                    }],
+                )
+                .unwrap()
+        };
+        PrepareResult::DrawReuse
+    }
+
+    fn draw(
+        &mut self,
+        layout: &B::PipelineLayout,
+        mut encoder: RenderPassEncoder<'_, B>,
+        index: usize,
+        _aux: &Aux<B>,
+    ) {
+        encoder.bind_graphics_descriptor_sets(
+            layout,
+            0,
+            Some(&self.sets[index]),
+            std::iter::empty(),
+        );
+        // This is a trick from Sascha Willems which uses just the gl_VertexIndex
+        // to calculate the position and uv coordinates for one full-scren "quad"
+        // which is actually just a triangle with two of the vertices positioned
+        // correctly off screen. This way we don't need a vertex buffer.
+        encoder.draw(0..3, 0..1);
+    }
+
+    fn dispose(mut self, factory: &mut Factory<B>, _aux: &mut Aux<B>) {
+        unsafe {
+            self.descriptor_pool.free_sets(self.sets.into_iter());
+            factory.destroy_descriptor_pool(self.descriptor_pool);
+        }
+    }
+}
